requirements.txt
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
caveclient==5.14.0
# Editable install with no version control (ccmodels==29.1.2024)
-e /Users/jacopobiggiogera/Desktop/con-con-models
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
datajoint==0.14.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
Flask==3.0.0
fonttools==4.47.2
idna==3.6
install==1.3.5
ipykernel==6.29.0
ipython==8.20.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
minio==7.2.3
nest-asyncio==1.5.9
networkx==2.6.3
numpy==1.26.3
otumat==0.3.1
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycparser==2.21
pycryptodome==3.20.0
pydot==2.0.0
Pygments==2.17.2
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.2
referencing==0.32.1
requests==2.31.0
rpds-py==0.17.1
scikit-learn==1.4.0
scipy==1.11.4
seaborn==0.13.1
six==1.16.0
stack-data==0.6.3
standard-transform==1.4.0
threadpoolctl==3.2.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
watchdog==3.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
